rename files to convention
